GPU: OpenCL doesn't support long long int

+++ GPU/GPUTracking/DataCompression/GPUTPCCompressionKernels.h
@@ -62,7 +62,7 @@ class GPUTPCCompressionKernels : public GPUKernelTemplate
 
   using Vec16 = unsigned short;
   using Vec32 = unsigned int;
-  using Vec64 = unsigned long long int;
+  using Vec64 = unsigned long int;
   using Vec128 = uint4;
 
   template <typename Scalar, typename BaseVector>
